requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
aiogram>=3.0.0
requests>=2.31.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
altair>=5.0.0
pyyaml>=6.0.0

